merge

+++ apollo.sh
@@ -16,56 +16,79 @@
 # limitations under the License.
 ###############################################################################
 
-
 #=================================================
 #                   Utils
 #=================================================
-DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-cd "${DIR}"
 
-source "${DIR}/scripts/apollo_base.sh"
+function source_apollo_base() {
+  DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+  cd "${DIR}"
 
-# the machine type, currently support x86_64, aarch64
-MACHINE_ARCH=$(uname -m)
+  source "${DIR}/scripts/apollo_base.sh"
+}
 
-IS_X86_64=false
-IS_AARCH64=false
+function apollo_check_system_config() {
+  # check operating system
+  OP_SYSTEM=$(uname -s)
+  case $OP_SYSTEM in
+    "Linux")
+      echo "System check passed. Build continue ..."
 
-BUILD_TARGETS=""
-TEST_TARGETS=""
+      # check system configuration
+      DEFAULT_MEM_SIZE="2.0"
+      MEM_SIZE=$(free | grep Mem | awk '{printf("%0.2f", $2 / 1024.0 / 1024.0)}')
+      if (( $(echo "$MEM_SIZE < $DEFAULT_MEM_SIZE" | bc -l) )); then
+         warning "System memory [${MEM_SIZE}G] is lower than minimum required memory size [2.0G]. Apollo build could fail."
+      fi
+      ;;
+    "Darwin")
+      warning "Mac OS is not officially supported in the current version. Build could fail. We recommend using Ubuntu 14.04."
+      ;;
+    *)
+      error "Unsupported system: ${OP_SYSTEM}."
+      error "Please use Linux, we recommend Ubuntu 14.04."
+      exit 1
+      ;;
+  esac
+}
 
-rm -rf ./third_party/ros
-if [ "$MACHINE_ARCH" == 'x86_64' ]; then
-   IS_X86_64=true
-elif [ "$MACHINE_ARCH" == 'aarch64' ]; then
-   IS_AARCH64=true
-fi
+function check_machine_arch() {
+  # the machine type, currently support x86_64, aarch64
+  MACHINE_ARCH=$(uname -m)
 
-if ! $IS_X86_64 && ! $IS_AARCH64 ; then
-   fail "Unknown machine architecture $MACHINE_ARCH"
-   exit 1
-else
-  ln -rs "$(pwd)/third_party/ros_$MACHINE_ARCH" "$(pwd)/third_party/ros"
-fi
+  # Generate WORKSPACE file based on marchine architecture
+  if [ "$MACHINE_ARCH" == 'x86_64' ]; then
+    sed "s/MACHINE_ARCH/x86_64/g" WORKSPACE.in > WORKSPACE
+  elif [ "$MACHINE_ARCH" == 'aarch64' ]; then
+    sed "s/MACHINE_ARCH/aarch64/g" WORKSPACE.in > WORKSPACE
+  else
+    fail "Unknown machine architecture $MACHINE_ARCH"
+    exit 1
+  fi
+}
 
 function check_esd_files() {
+  CAN_CARD="fake_can"
+
   if [ -f ./third_party/can_card_library/esd_can/include/ntcan.h \
       -a -f ./third_party/can_card_library/esd_can/lib/libntcan.so \
       -a -f ./third_party/can_card_library/esd_can/lib/libntcan.so.4 \
       -a -f ./third_party/can_card_library/esd_can/lib/libntcan.so.4.0.1 ]; then
       USE_ESD_CAN=true
+      CAN_CARD="esd_can"
   else
-      warning "${YELLOW}ESD CAN library supplied by ESD Electronics does not exit.${NO_COLOR}"
-      warning "${YELLOW}If you need ESD CAN, please refer to third_party/can_card_library/esd_can/README.md${NO_COLOR}"
+      warning "ESD CAN library supplied by ESD Electronics does not exist. If you need ESD CAN, please refer to third_party/can_card_library/esd_can/README.md."
       USE_ESD_CAN=false
   fi
 }
 
-check_esd_files
-
 function generate_build_targets() {
   BUILD_TARGETS=$(bazel query //... | grep -v "_test$" | grep -v "third_party" \
     | grep -v "_cpplint$" | grep -v "release" | grep -v "kernel")
+  if [ $? -ne 0 ]; then
+    fail 'Build failed!'
+  fi
+
   if ! $USE_ESD_CAN; then
      BUILD_TARGETS=$(echo $BUILD_TARGETS |tr ' ' '\n' | grep -v "hwmonitor" | grep -v "esd")
   fi
@@ -73,6 +96,10 @@ function generate_build_targets() {
 
 function generate_test_targets() {
   TEST_TARGETS=$(bazel query //... | grep "_test$" | grep -v "third_party" | grep -v "kernel")
+  if [ $? -ne 0 ]; then
+    fail 'Test failed!'
+  fi
+
   if ! $USE_ESD_CAN; then
      TEST_TARGETS=$(echo $TEST_TARGETS| tr ' ' '\n' | grep -v "hwmonitor" | grep -v "esd")
   fi
@@ -82,18 +109,14 @@ function generate_test_targets() {
 #              Build functions
 #=================================================
 
-
 function apollo_build() {
-  START_TIME=$(($(date +%s%N)/1000000))
+  START_TIME=$(get_now)
 
-  # Avoid query release directory.
-  if [ -d release ];then
-    rm -rf release
-  fi
+  echo "Start building, please wait ..."
   generate_build_targets
   echo "Building on $MACHINE_ARCH, with targets:"
   echo "$BUILD_TARGETS"
-  echo "$BUILD_TARGETS" | xargs bazel --batch --batch_cpu_scheduling build --define ARCH="$MACHINE_ARCH" --define CAN_CARD=${CAN_CARD} --cxxopt=-DUSE_ESD_CAN=${USE_ESD_CAN} -c dbg
+  echo "$BUILD_TARGETS" | xargs bazel --batch --batch_cpu_scheduling build --jobs=10 --define ARCH="$MACHINE_ARCH" --define CAN_CARD=${CAN_CARD} --cxxopt=-DUSE_ESD_CAN=${USE_ESD_CAN} -c dbg
   if [ $? -eq 0 ]; then
     success 'Build passed!'
   else
@@ -103,7 +126,7 @@ function apollo_build() {
 }
 
 function check() {
-  local check_start_time=$START_TIME
+  local check_start_time=$(get_now)
   apollo_build && run_test && run_lint
 
   START_TIME=$check_start_time
@@ -116,25 +139,23 @@ function check() {
   fi
 }
 
-# print the given message in red txt with white background.
-function echo_red_white() {
-  echo -e "$(tput setaf 1)$(tput setab 7)"$1 "\e[0m"
-}
-
 function warn_proprietary_sw() {
-  echo_red_white "The release built contains proprietary software provided by other parties. "
-  echo_red_white "Make sure you have obtained proper licensing agreement for redistribution "
-  echo_red_white "if you intend to publish the release package built. "
-  echo_red_white "Such licensing agreement is solely between you and the other parties, "
-  echo_red_white "and is NOT covered by the license terms of the Apollo project "
-  echo_red_white "(see file LICENSE)."
+  echo -e "${RED}The release built contains proprietary software provided by other parties.${NO_COLOR}"
+  echo -e "${RED}Make sure you have obtained proper licensing agreement for redistribution${NO_COLOR}"
+  echo -e "${RED}if you intend to publish the release package built.${NO_COLOR}"
+  echo -e "${RED}Such licensing agreement is solely between you and the other parties,${NO_COLOR}"
+  echo -e "${RED}and is not covered by the license terms of the apollo project${NO_COLOR}"
+  echo -e "${RED}(see file license).${NO_COLOR}"
 }
 
 function release() {
   ROOT_DIR=$HOME/.cache/release
-  rm -rf $ROOT_DIR
+  if [ -d $ROOT_DIR ];then
+    rm -rf $ROOT_DIR
+  fi
+  mkdir -p $ROOT_DIR
 
-  #modules
+  # modules
   MODULES_DIR=$ROOT_DIR/modules
   mkdir -p $MODULES_DIR
   for m in control canbus localization decision perception prediction planning
@@ -149,29 +170,36 @@ function release() {
         cp -r modules/$m/conf $TARGET_DIR
     fi
   done
-  #control tools
+
+  # control tools
   mkdir $MODULES_DIR/control/tools
-  cp bazel-bin/modules/control/tools/enter_auto_mode $MODULES_DIR/control/tools
   cp bazel-bin/modules/control/tools/pad_terminal $MODULES_DIR/control/tools
+
   #remove all pyc file in modules/
   find modules/ -name "*.pyc" | xargs -I {} rm {}
   cp -r modules/tools $MODULES_DIR
-  #ros
-  cp -Lr third_party/ros $ROOT_DIR/
-  #scripts
+
+  # ros
+  cp -Lr bazel-apollo/external/ros $ROOT_DIR/
+
+  # scripts
   cp -r scripts $ROOT_DIR
+
   #dreamview
   cp -Lr bazel-bin/modules/dreamview/dreamview.runfiles/apollo/modules/dreamview $MODULES_DIR
   cp -r modules/dreamview/conf $MODULES_DIR/dreamview
-  #common data
+
+  # common data
   mkdir $MODULES_DIR/common
   cp -r modules/common/data $MODULES_DIR/common
-  #hmi
+
+  # hmi
   mkdir -p $MODULES_DIR/hmi/ros_node $MODULES_DIR/hmi/utils
   cp bazel-bin/modules/hmi/ros_node/ros_node_service $MODULES_DIR/hmi/ros_node/
   cp -r modules/hmi/conf $MODULES_DIR/hmi
   cp -r modules/hmi/web $MODULES_DIR/hmi
   cp -r modules/hmi/utils/*.py $MODULES_DIR/hmi/utils
+
   # lib
   LIB_DIR=$ROOT_DIR/lib
   mkdir $LIB_DIR
@@ -181,7 +209,7 @@ function release() {
     do
         cp third_party/can_card_library/$m/lib/* $LIB_DIR
     done
-    #hw check
+    # hw check
     mkdir -p $MODULES_DIR/monitor/hwmonitor/hw_check/
     cp bazel-bin/modules/monitor/hwmonitor/hw_check/can_check $MODULES_DIR/monitor/hwmonitor/hw_check/
     cp bazel-bin/modules/monitor/hwmonitor/hw_check/gps_check $MODULES_DIR/monitor/hwmonitor/hw_check/
@@ -189,21 +217,21 @@ function release() {
     cp bazel-bin/modules/monitor/hwmonitor/hw/tools/esdcan_test_app $MODULES_DIR/monitor/hwmonitor/hw/tools/
   fi
   cp -r bazel-genfiles/* $LIB_DIR
+
   # doc
   cp -r docs $ROOT_DIR
   cp LICENSE $ROOT_DIR
   cp third_party/ACKNOWLEDGEMENT.txt $ROOT_DIR
-  #release info
+
+  # release info
   META=${ROOT_DIR}/meta.txt
   echo "Git commit: $(git show --oneline  -s | awk '{print $1}')" > $META
   echo "Build time: $TIME" >>  $META
 }
 
 function gen_coverage() {
-  # Avoid query release directory.
-  if [ -d release ];then
-    rm -rf release
-  fi
+  START_TIME=$(get_now)
+
   bazel clean
   generate_test_targets
   echo "$TEST_TARGETS" | xargs bazel test --define ARCH="$(uname -m)" --define CAN_CARD=${CAN_CARD} --cxxopt=-DUSE_ESD_CAN=${USE_ESD_CAN} -c dbg --config=coverage
@@ -230,18 +258,13 @@ function gen_coverage() {
       "tools/*" \
       -o $COV_DIR/stripped_conv.info
   genhtml $COV_DIR/stripped_conv.info --output-directory $COV_DIR/report
-  echo "Generated coverage report in $COV_DIR/report/index.html"
+
+  success 'Generated coverage report in $COV_DIR/report/index.html'
 }
 
 function run_test() {
-  START_TIME=$(($(date +%s%N)/1000000))
+  START_TIME=$(get_now)
 
-  # Avoid query release directory.
-  if [ -d release ];then
-    rm -rf release
-  fi
-  # FIXME(all): when all unit test passed, switch back.
-  # bazel test --config=unit_test -c dbg //...
   generate_test_targets
   echo "$TEST_TARGETS" | xargs bazel test --define "ARCH=$MACHINE_ARCH"  --define CAN_CARD=${CAN_CARD} --config=unit_test --cxxopt=-DUSE_ESD_CAN=${USE_ESD_CAN} -c dbg --test_verbose_timeout_warnings
   if [ $? -eq 0 ]; then
@@ -258,12 +281,21 @@ function run_cpp_lint() {
 }
 
 function run_bash_lint() {
-  FILES=$(find "${DIR}" -type f -name "*.sh" | grep -v ros | grep -v kernel)
+  FILES=$(find "${APOLLO_ROOT_DIR}" -type f -name "*.sh" | grep -v ros | grep -v kernel)
   echo "${FILES}" | xargs shellcheck
 }
 
 function run_lint() {
-  START_TIME=$(($(date +%s%N)/1000000))
+  START_TIME=$(get_now)
+
+  # Add cpplint rule to BUILD files that do not contain it.
+  for file in $(find modules -name BUILD | \
+    xargs grep -l -E 'cc_library|cc_test|cc_binary' | xargs grep -L 'cpplint()')
+  do
+    sed -i '1i\load("//tools:cpplint.bzl", "cpplint")\n' $file
+    sed -i -e '$a\\ncpplint()' $file
+  done
+
   run_cpp_lint
 
   if [ $? -eq 0 ]; then
@@ -274,21 +306,22 @@ function run_lint() {
 }
 
 function clean() {
-   bazel clean --async
+  bazel clean --async
 }
 
 function buildify() {
-   wget \
-   https://github.com/bazelbuild/buildtools/releases/download/0.4.5/buildifier \
-   -O ~/.buildifier
-   chmod +x ~/.buildifier
-   find . -name BUILD -type f -exec ~/.buildifier -showlog -mode=fix {} +
-   if [ $? -eq 0 ]; then
-     success 'Buildify worked!'
-   else
-     fail 'Buildify failed!'
-   fi
-   rm ~/.buildifier
+  START_TIME=$(get_now)
+
+  local buildifier_url=https://github.com/bazelbuild/buildtools/releases/download/0.4.5/buildifier
+  wget $buildifier_url -O ~/.buildifier
+  chmod +x ~/.buildifier
+  find . -name BUILD -type f -exec ~/.buildifier -showlog -mode=fix {} +
+  if [ $? -eq 0 ]; then
+    success 'Buildify worked!'
+  else
+    fail 'Buildify failed!'
+  fi
+  rm ~/.buildifier
 }
 
 function print_usage() {
@@ -299,6 +332,7 @@ function print_usage() {
   buildify: fix style of BUILD files
   check: run build/lint/test, please make sure it passes before checking in new code
   clean: runs Bazel clean
+  config: run configurator tool
   coverage: generate test coverage report
   doc: generate doxygen document
   lint: run code style check
@@ -321,85 +355,105 @@ function version() {
   echo "Date: ${date}"
 }
 
-function buildgnss() {
-    CURRENT_PATH=$(pwd)
-    MACHINE_ARCH="$(uname -m)"
-    INSTALL_PATH="${CURRENT_PATH}/third_party/ros_${MACHINE_ARCH}"
+function build_gnss() {
+  CURRENT_PATH=$(pwd)
+  if [ -d "${CURRENT_PATH}/bazel-apollo/external/ros" ]; then
+    ROS_PATH="${CURRENT_PATH}/bazel-apollo/external/ros"
+  else
+    warning "ROS not found. Run apolllo.sh build first."
+    exit 1
+  fi
 
-    source "${INSTALL_PATH}/setup.bash"
+  source "${ROS_PATH}/setup.bash"
 
-    protoc modules/common/proto/error_code.proto --cpp_out=./
-    protoc modules/common/proto/header.proto --cpp_out=./
-    protoc modules/common/proto/geometry.proto --cpp_out=./
+  protoc modules/common/proto/error_code.proto --cpp_out=./
+  protoc modules/common/proto/header.proto --cpp_out=./
+  protoc modules/common/proto/geometry.proto --cpp_out=./
 
-    protoc modules/localization/proto/imu.proto --cpp_out=./
-    protoc modules/localization/proto/gps.proto --cpp_out=./
-    protoc modules/localization/proto/pose.proto --cpp_out=./
+  protoc modules/localization/proto/imu.proto --cpp_out=./
+  protoc modules/localization/proto/gps.proto --cpp_out=./
+  protoc modules/localization/proto/pose.proto --cpp_out=./
 
-    protoc modules/drivers/gnss/proto/gnss.proto --cpp_out=./
-    protoc modules/drivers/gnss/proto/imu.proto --cpp_out=./
-    protoc modules/drivers/gnss/proto/ins.proto --cpp_out=./
-    protoc modules/drivers/gnss/proto/config.proto --cpp_out=./
-    protoc modules/drivers/gnss/proto/gnss_status.proto --cpp_out=./
-    protoc modules/drivers/gnss/proto/gpgga.proto --cpp_out=./
+  protoc modules/drivers/gnss/proto/gnss.proto --cpp_out=./
+  protoc modules/drivers/gnss/proto/imu.proto --cpp_out=./
+  protoc modules/drivers/gnss/proto/ins.proto --cpp_out=./
+  protoc modules/drivers/gnss/proto/config.proto --cpp_out=./
+  protoc modules/drivers/gnss/proto/gnss_status.proto --cpp_out=./
+  protoc modules/drivers/gnss/proto/gpgga.proto --cpp_out=./
 
-    cd modules
-    catkin_make_isolated --install --source drivers \
-        --install-space "${INSTALL_PATH}" -DCMAKE_BUILD_TYPE=Release \
-        --cmake-args --no-warn-unused-cli
-    find "${INSTALL_PATH}" -name "*.pyc" -print0 | xargs -0 rm -rf
-    cd -
+  cd modules
+  catkin_make_isolated --install --source drivers \
+    --install-space "${ROS_PATH}" -DCMAKE_BUILD_TYPE=Release \
+    --cmake-args --no-warn-unused-cli
+  find "${ROS_PATH}" -name "*.pyc" -print0 | xargs -0 rm -rf
+  cd -
 
-    rm -rf modules/common/proto/*.pb.cc
-    rm -rf modules/common/proto/*.pb.h
-    rm -rf modules/drivers/gnss/proto/*.pb.cc
-    rm -rf modules/drivers/gnss/proto/*.pb.h
-    rm -rf modules/localization/proto/*.pb.cc
-    rm -rf modules/localization/proto/*.pb.h
+  rm -rf modules/common/proto/*.pb.cc
+  rm -rf modules/common/proto/*.pb.h
+  rm -rf modules/drivers/gnss/proto/*.pb.cc
+  rm -rf modules/drivers/gnss/proto/*.pb.h
+  rm -rf modules/localization/proto/*.pb.cc
+  rm -rf modules/localization/proto/*.pb.h
 
-    rm -rf modules/.catkin_workspace
-    rm -rf modules/build_isolated/
-    rm -rf modules/devel_isolated/
+  rm -rf modules/.catkin_workspace
+  rm -rf modules/build_isolated/
+  rm -rf modules/devel_isolated/
 }
 
-case $1 in
-  check)
-    check
-    ;;
-  build)
-    apollo_build
-    ;;
-  buildify)
-    buildify
-    ;;
-  buildgnss)
-    buildgnss
-    ;;
-  doc)
-    gen_doc
-    ;;
-  lint)
-    run_lint
-    ;;
-  test)
-    run_test
-    ;;
-  release)
-    release 1
-    ;;
-  release_noproprietary)
-    release 0
-    ;;
-  coverage)
-    gen_coverage
-    ;;
-  clean)
-    clean
-    ;;
-  version)
-    version
-    ;;
-  *)
-    print_usage
-    ;;
-esac
+function config() {
+  ${APOLLO_ROOT_DIR}/scripts/configurator.sh
+}
+
+function main() {
+  source_apollo_base
+  apollo_check_system_config
+  check_machine_arch
+  check_esd_files
+
+  case $1 in
+    check)
+      check
+      ;;
+    build)
+      apollo_build
+      ;;
+    buildify)
+      buildify
+      ;;
+    buildgnss)
+      build_gnss
+      ;;
+    config)
+      config
+      ;;
+    doc)
+      gen_doc
+      ;;
+    lint)
+      run_lint
+      ;;
+    test)
+      run_test
+      ;;
+    release)
+      release 1
+      ;;
+    release_noproprietary)
+      release 0
+      ;;
+    coverage)
+      gen_coverage
+      ;;
+    clean)
+      clean
+      ;;
+    version)
+      version
+      ;;
+    *)
+      print_usage
+      ;;
+  esac
+}
+
+main $@
